feat: Complete FastNoise library port
This commit completes the C# port of the FastNoise library.

- Implemented all noise types: Simplex, Perlin, Cubic, Cellular, and Value.
- Added fractal variations for all noise types: FBM, Billow, and RigidMulti.
- Refactored the codebase to be more object-oriented and maintainable.
- Replaced the single `xFastNoise.cs` file with a clean, one-class-per-file structure.
- Changed all floating-point calculations from `float` to `double` for improved precision.
- Made cellular jitter configurable via the `INoiseSettings` interface.
- Made seed mutation in fractal classes exception-safe by wrapping it in a `try...finally` block.

+++ FastNoise/Noises/INoise.cs
@@ -2,7 +2,7 @@
 {
     public interface INoise
     {
-        float GetNoise(Vector2 vec);
-        float GetNoise(Vector3 vec);
+        double GetNoise(Vector2 vec);
+        double GetNoise(Vector3 vec);
     }
 }
